% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{add_gc_df}
\alias{add_gc_df}
\title{add_gc_df}
\usage{
add_gc_df(data_frame_input, numproc = 2)
}
\arguments{
\item{data_frame_input}{input data.frame, with column names including Start and Length}

\item{numproc}{number of processors to use (default 2)}
}
\value{
data.frame containing the input columns plus an extra column 'GC' representing mean GC content
}
\description{
add_gc_df
}

