% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{append_colors}
\alias{append_colors}
\title{append_colors}
\usage{
append_colors(df_in, classif_df, type = "coverage",
  exceptional_regions = NULL, exceptional_regions_col = "orange")
}
\arguments{
\item{df_in}{input data.frame}

\item{classif_df}{data.frame of classifications}

\item{type}{'coverage' or 'zygosity'}

\item{exceptional_regions}{data.frame containing exceptional regions (Chr, Start, End)}

\item{exceptional_regions_col}{color to use for exceptional regions, default orange}
}
\value{
data.frame with additional column 'Color' (navyblue, blue, green, black, red2, red4)
}
\description{
Append colors to a data.frame of classifications
}

