% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{append_covzyg_predictions}
\alias{append_covzyg_predictions}
\title{append_covzyg_predictions}
\usage{
append_covzyg_predictions(classifs, tc, C1, Z1)
}
\arguments{
\item{classifs}{data.frame of classifications}

\item{tc}{Inferred tumor content}

\item{C1}{coverage data}

\item{Z1}{zygosity data}
}
\value{
a list with components coverage and zygosity
}
\description{
Append predicted values to coverage and zygosity data
}

