% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{bed2df}
\alias{bed2df}
\title{bed2df}
\usage{
bed2df(bedfile, add_region_names = TRUE)
}
\arguments{
\item{bedfile}{path to bedfile (can contain optional header, ignored if it exists.)}

\item{add_region_names}{if TRUE, adds 1-based region names in the form chr:start-end}
}
\value{
data.frame containing columns Chr, Start, End, and optionally Region for targets on 1-22,X,Y
}
\description{
bed2df
}

