% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna_pipeline.R
\name{cna_pipeline}
\alias{cna_pipeline}
\title{cna_pipeline}
\usage{
cna_pipeline(BAMT = "", BAMN = "", VCF = "", POPFREQ_VCF = "",
  VCF_FIELDS = "AO:DP:Tumor:Normal:MAF", PREFIX = "", OUTDIR = "/out",
  BEDFILE = "", NP = 2, COVERAGE_INPUT = "bam", ZYGOSITY_INPUT = "vcf",
  COVERAGE_FILE = "", ZYGOSITY_FILE = "", NORMALPOOLFILE = NULL,
  POLYMORPHIC_REGIONS = NULL, LOCUS_ANNOTATION_FILE = NULL,
  EXCLUDE_LOCI = TRUE, GENDER = "infer", config = "")
}
\arguments{
\item{BAMT}{Tumor Bam File (required if COVERAGE_INPUT = 'bam')}

\item{BAMN}{Normal Bam File (required if COVERAGE_INPUT = 'bam')}

\item{VCF}{VCF file containing variants (required if ZYGOSITY_INPUT = 'vcf')}

\item{POPFREQ_VCF}{VCF file containing population frequencies (optional but recommended; used only if ZYGOSITY_INPUT = 'vcf')}

\item{VCF_FIELDS}{A colon-separated text string encoding relevant fields of the two VCF files, default "AO:DP:Tumor:Normal:MAF".
In order, these reflect the fields giving variant read counts, total depth of coverage, name of the tumor sample, name of the normal sample,
and the field giving population frequencies.  The first four come from VCF, the fifth comes from POPFREQ_VCF.}

\item{PREFIX}{Sample prefix used to label all output files (required)}

\item{OUTDIR}{Output directory (/out by default)}

\item{BEDFILE}{Path to bedfile containing target regions.}

\item{NP}{Number of processors to use.}

\item{COVERAGE_INPUT}{"bam" or "txt", if "bam" then specify BAMT and BAMN, if "txt" then specify COVERAGE_FILE}

\item{ZYGOSITY_INPUT}{"vcf" or "txt", if "vcf" then specify VCF, if "txt" then specify ZYGOSITY_FILE}

\item{COVERAGE_FILE}{path to text file (with header row) containing coverage quantifications per targeted region.
Required if COVERAGE_INPUT = "txt".  Required column names are Chr, Start, End, MeanCovT, and MeanCovN.
Chromosome names must be in {1,...,22,'X','Y'}, and the Chr/Start/End columns should exactly match the lines of BEDFILE.}

\item{ZYGOSITY_FILE}{path to text file (with header row) containing minimally-filtered variants, their read counts in tumor and normal,
and their population frequencies.  This file should include all detected heterozygous SNPs from the normal sample but can include
homozygous SNPS, artifacts, etc. as the pipeline does its own filtering.  Required if ZYGOSITY_INPUT = "txt".
Required column names are Chr, Pos, Ref, Alt, Var.T, Tot.T, Var.N, and Tot.N.
Ref.N and Ref.T may also be specified, otherwise they are assumed to be Tot.N-Var.N and Tot.T-Var.T, respectively.
An additional column Pop_Freq containing 1000 genomes or ExAC population frequencies (values between 0 and 1) is strongly recommended,
otherwise every variant is assumed to be a common SNP.  Chromosome names must be in {1,...,22,'X','Y'}.}

\item{NORMALPOOLFILE}{RDS file containing normal pool quantifications for the capture platform of interest}

\item{POLYMORPHIC_REGIONS}{text file containing a set of regions to mask because they are highly polymorphic}

\item{LOCUS_ANNOTATION_FILE}{text file with columns Locus,Chr,Start,End against which to annotate results}

\item{EXCLUDE_LOCI}{whether to exclude certain loci (exclude_loci_hg19) from classification, default TRUE}

\item{GENDER}{either 'male', 'female', or 'infer' (default = 'infer' which infers the gender from the data)}

\item{config}{path to text file containing non-default values of parameters}
}
\description{
run the full CNA pipeline
}

