% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{compute_fit}
\alias{compute_fit}
\title{compute_fit}
\usage{
compute_fit(zygosity, coverage, coverage_colname, exceptional_regions = NULL)
}
\arguments{
\item{zygosity}{data.frame containing zygosity data}

\item{coverage}{data.frame containing coverage data}

\item{coverage_colname}{column name to use for coverage}

\item{exceptional_regions}{regions to exclude from computation of fit statistics}
}
\value{
vector of length two containing ratios of zygosity and coverage data deviations scaled to their theoretical lower limits
}
\description{
Compute the fit of estimated zygosity and coverage values
}

