% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_functions.R
\name{coverage_only_segmentation}
\alias{coverage_only_segmentation}
\title{coverage_only_segmentation}
\usage{
coverage_only_segmentation(coverage_data, coverage_colname, chrinfo)
}
\arguments{
\item{coverage_data}{coverage data.frame}

\item{coverage_colname}{column name to use for coverage}

\item{chrinfo}{chrinfo data.frame}
}
\value{
data.frame containing segmentation
}
\description{
Germline segmentation
}

