% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_functions.R
\name{create.statelist}
\alias{create.statelist}
\title{create.statelist}
\usage{
create.statelist(theta, copyrange = 0:4, lr.mean = 0, vf.mean = 0.5)
}
\arguments{
\item{theta}{Tumor purity value (between 0 and 1)}

\item{copyrange}{Vector of integer copy number values to consider (default 0:4)}

\item{lr.mean}{Log2 coverage ratio mean (default 0)}

\item{vf.mean}{Variant fraction mean (default 0)}
}
\value{
data.frame containing columns State, Copies, Num.Var, Normal.Genotype, Tumor.Genotype, LR, and VF
}
\description{
Create a list of copy number/zygosity states for a given tumor purity value
}

