% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_functions.R
\name{e.step}
\alias{e.step}
\title{e.step}
\usage{
e.step(theta, segs, snplist, coveragelist, coverage_colname, stdev,
  copyrange = 0:4, lr.mean = 0, vf.mean = 0.5)
}
\arguments{
\item{theta}{Tumor content estimate}

\item{segs}{data.frame containing segments}

\item{snplist}{List of snp data per segment, in the same order as segs data.frame}

\item{coveragelist}{List of coverage data per segment, in the same order as segs data.frame}

\item{coverage_colname}{Column name to use from coverage data.frame}

\item{stdev}{Global standard deviation of log2 coverage ratios}

\item{copyrange}{Vector of integer copy number states to consider (default 0:4)}

\item{lr.mean}{Global Log2 coverage ratio mean (default 0)}

\item{vf.mean}{Global Variant Fraction mean (default 0.5)}
}
\value{
a list with two components, Probs and States
}
\description{
Perform E-step of EM algorithm
}

