% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exclude_loci_hg19}
\alias{exclude_loci_hg19}
\title{Loci to Exclude (hg19 coordinates)}
\format{A data frame}
\usage{
data(exclude_loci_hg19)
}
\description{
A data.frame containing hg19 regions to exclude from classification.
}
\keyword{datasets}

