% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{filter_to_plausible_adjustments}
\alias{filter_to_plausible_adjustments}
\title{filter_to_plausible_adjustments}
\usage{
filter_to_plausible_adjustments(fits.df)
}
\arguments{
\item{fits.df}{data.frame of fits}
}
\value{
vector of plausible adjustments
}
\description{
Filter a set of possible adjustments to the most plausible ones based on the shape of the adjustment/fit function
}

