% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{fit_models}
\alias{fit_models}
\title{fit_models}
\usage{
fit_models(adj_df, coverage_data, coverage_data_colname, seg_data,
  zygosity_data, numproc, chrinfo, exceptional_regions = NULL,
  gender = "female")
}
\arguments{
\item{adj_df}{data.frame of adjustments to consider}

\item{coverage_data}{coverage data.frame}

\item{coverage_data_colname}{colname to use for coverage}

\item{seg_data}{segmented data}

\item{zygosity_data}{zygosity data.frame}

\item{numproc}{Number of processes to use}

\item{chrinfo}{data.frame containing chromosome information}

\item{exceptional_regions}{regions to exclude when computing fit statistics}

\item{gender}{gender of sample (default female)}
}
\value{
a list, one element per ploidy adjustment, containing such information as the overall classifications, ploidy, and goodness of fit
}
\description{
Fit classification models across a range of ploidy adjustments
}

