% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_functions.R
\name{germline_classification}
\alias{germline_classification}
\title{germline_classification}
\usage{
germline_classification(covdata, segdata, outdir, prefix, chrinfo,
  gene_df = NULL)
}
\arguments{
\item{covdata}{coverage data.frame}

\item{segdata}{data.frame of segments}

\item{outdir}{output directory (results go to 'germline' subdirectory of outdir)}

\item{prefix}{sample prefix to append to beginning of each file}

\item{chrinfo}{chromosome information as generated by get_chrinfo function}

\item{gene_df}{data.frame containing genes with which to annotate output files (default NULL, in which case results are not annotated)}
}
\description{
Classification of germline variants
}

