% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_functions.R
\name{get.fb.lik}
\alias{get.fb.lik}
\title{get.fb.lik}
\usage{
get.fb.lik(x, n, p)
}
\arguments{
\item{x}{Number of variant reads (binomial number of successes)}

\item{n}{Amount of coverage (binomial number of experiments)}

\item{p}{Zygosity mean (binomial probability)}
}
\value{
Likelihood using the beta-binomial function
}
\description{
Retrieve likelihood for zygosity data
}

