% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{get_chrinfo}
\alias{get_chrinfo}
\title{get_chrinfo}
\usage{
get_chrinfo(chrnames, numproc = 2)
}
\arguments{
\item{chrnames}{Vector of chromosome names to use (default 1:22,X,Y)}

\item{numproc}{Number of processors to use}
}
\value{
data.frame containing ChrName, Length_Orig (in base pairs), Length, and PrevLengths
}
\description{
get_chrinfo
}

