% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{get_classifications_from_models}
\alias{get_classifications_from_models}
\title{get_classifications_from_models}
\usage{
get_classifications_from_models(models, adj_df, cov_df, seg_df, colname,
  numproc)
}
\arguments{
\item{models}{list of models}

\item{adj_df}{data.frame of adjustments}

\item{cov_df}{data.frame containing unadjusted coverage values}

\item{seg_df}{data.frame containing unadjusted segmentation values}

\item{colname}{column name to use}

\item{numproc}{number of processors to use}
}
\value{
list of classifications, one per model
}
\description{
retrieve classifications from a set of models
}

