% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{get_gc}
\alias{get_gc}
\title{get_gc}
\usage{
get_gc(startvec, lengthvec, chrom, numproc = 2)
}
\arguments{
\item{startvec}{vector of start positions}

\item{lengthvec}{vector of exon lengths}

\item{chrom}{chromosome name (from 1-22,X,Y)}

\item{numproc}{number of processors (default 2) to use}
}
\value{
numeric vector containing mean GC contents for the regions specified
}
\description{
get_gc
}

