% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{get_max_readlength}
\alias{get_max_readlength}
\title{get_max_readlength}
\usage{
get_max_readlength(bamfile, N = 10000)
}
\arguments{
\item{bamfile}{path to bam file}

\item{N}{number of reads to use}
}
\value{
numeric value representing maximum read length (NULL if bamfile doesn't exist)
}
\description{
get_max_readlength
}

