% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{get_model_info}
\alias{get_model_info}
\title{get_model_info}
\usage{
get_model_info(models, adj_df, coverage_df, seg_data, numproc, colname)
}
\arguments{
\item{models}{list of models}

\item{adj_df}{data.frame containing adjustment information}

\item{coverage_df}{data.frame containing coverage data}

\item{seg_data}{data.frame containing segmented data}

\item{numproc}{number of processors to use}

\item{colname}{column name to use}
}
\value{
list of classifications and summary stats
}
\description{
Get information for a set of models
}

