% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_plotting_coords}
\alias{get_plotting_coords}
\title{get_plotting_coords}
\usage{
get_plotting_coords(position_vec, chr_vec, numproc = 2)
}
\arguments{
\item{position_vec}{vector containing chromosomal positions}

\item{chr_vec}{vector containing chromosome names (from 1-22,X,Y)}

\item{numproc}{number of processors to use (default 2)}
}
\value{
a vector of correctly-spaced coordinates to use for plotting along the whole genome, in units of 10kb
}
\description{
get_plotting_coords
}

