% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{get_quant_mx}
\alias{get_quant_mx}
\title{get_quant_mx}
\usage{
get_quant_mx(input_vec, ref_mx, gc_vec, mincov = 30)
}
\arguments{
\item{input_vec}{vector containing coverage quantifications for the sample of interest}

\item{ref_mx}{matrix (rows = targets, columns = samples) containing quantifications for the reference samples}

\item{gc_vec}{vector of mean GC content for the targets}

\item{mincov}{minimum reference coverage to use for quantification}
}
\value{
matrix of corrected values (in the same order as the input reference matrix)
}
\description{
Quantify a sample against a matrix of reference samples
}

