% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{get_tc_lower_limit}
\alias{get_tc_lower_limit}
\title{get_tc_lower_limit}
\usage{
get_tc_lower_limit(coverage_df, colname_LR, colname_cov)
}
\arguments{
\item{coverage_df}{data.frame of coverage values}

\item{colname_LR}{column name containing log-ratio values to use}

\item{colname_cov}{column name containing coverage value to use as a minimum cutoff}
}
\value{
vector containing lower limit for tumor content estimation, standard deviation, and minimum sequencing coverage
}
\description{
Compute lower limit for tumor content estimation
}

