% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{getFeatureCounts}
\alias{getFeatureCounts}
\title{getFeatureCounts}
\usage{
getFeatureCounts(bamfile, bedfile, padding = 0, threads = 1,
  writedir = NULL)
}
\arguments{
\item{bamfile}{full path to bam file}

\item{bedfile}{full path to bed file}

\item{padding}{number of bases (default 0) on each side of target on which to quantify reads}

\item{threads}{number of threads to use (default 1)}
}
\value{
data.frame containing per-target counts from the bam file for each target in the bed file
}
\description{
getFeatureCounts
}

