% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{infer_gender}
\alias{infer_gender}
\title{infer_gender}
\usage{
infer_gender(coverage_vec, chrom_vec)
}
\arguments{
\item{coverage_vec}{vector containing coverage values across all target regions}

\item{chrom_vec}{vector of chromosomes (from 1-22,X,Y)}
}
\value{
character vector containing inferred gender ("female" or "male")
}
\description{
Infers gender based on Y-chromosome coverage
}

