% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{initialize_adjustments}
\alias{initialize_adjustments}
\title{initialize_adjustments}
\usage{
initialize_adjustments(adjustment_range = seq(-0.6, 1, by = 0.02),
  full_search = FALSE, seg_data = NULL, tc_lower = 0.3)
}
\arguments{
\item{adjustment_range}{vector of values to consider for ploidy adjustment}

\item{full_search}{if TRUE, scans all values in adjustment_range (slow), if FALSE, limits to likely candidates}

\item{seg_data}{segmentation data.frame (default NULL); required if full_search is FALSE}

\item{tc_lower}{lower limit of tumor content to consider}
}
\value{
data.frame containing Adjustment and TC_Lower
}
\description{
Initialize adjustments for ploidy correction
}

