% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{integrated_classification}
\alias{integrated_classification}
\title{integrated_classification}
\usage{
integrated_classification(covdata, zygdata, segdata, outdir, refsample,
  prefix = "Sample", numproc = 2, locus_annotation_file = NULL,
  exceptional_regions = NULL, gender = "female")
}
\arguments{
\item{covdata}{data.frame containing coverage}

\item{zygdata}{data.frame containing zygosity}

\item{segdata}{data.frame containing segmentation results}

\item{outdir}{output directory}

\item{refsample}{reference sample ('normal' or 'pool')}

\item{prefix}{sample prefix to use for result files}

\item{numproc}{Number of processors to use}

\item{locus_annotation_file}{data.frame containing gene structure for annotation purposes}

\item{exceptional_regions}{regions not to classify}
}
\description{
Ploidy-adjusted classification across a set of models
}

