% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_functions.R
\name{integrated_segmentation}
\alias{integrated_segmentation}
\title{integrated_segmentation}
\usage{
integrated_segmentation(coverage_data, zygosity_data, coverage_colname,
  use_normal_sample = TRUE, exceptional_regions = NULL)
}
\arguments{
\item{coverage_data}{data.frame containing coverage quantifications}

\item{zygosity_data}{data.frame containing variants}

\item{coverage_colname}{column name of coverage_data to use as log2 coverage ratios}

\item{use_normal_sample}{if TRUE, uses normal sample; if FALSE uses only tumor sample}

\item{exceptional_regions}{set of non-overlapping regions, each of which is its own segment.  data.frame with columns Chr, Start, End.  default = NULL.}
}
\value{
data.frame of segments with columns Segment, Chr, Start, End, Num.Agg.SNPs, Dev.Zyg.Seg, Num.Exons, Log2.Coverage.Ratio, Num.SNPs, Num.Exons.NM
}
\description{
Joint segmentation of coverage and zygosity data.
}

