% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{label_suspicious_adjustments}
\alias{label_suspicious_adjustments}
\title{label_suspicious_adjustments}
\usage{
label_suspicious_adjustments(df_in, max_homozygous_prop_genome = 0.05,
  max_exons_seg_homozygous = 200, max_avg_zyg_shift_neutral = 0.1,
  min_avg_zyg_shift_onecopy = 0.1)
}
\arguments{
\item{df_in}{input data.frame containing summary statistics for a set of adjustments}

\item{max_homozygous_prop_genome}{Maximum proportion of the genome called as homozygously deleted before an adjustment is tagged as 'very suspicious' (default 0.05)}

\item{max_exons_seg_homozygous}{Maximum number of exons in a homozygously deleted segment before an adjustment is tagged as 'very suspicious' (default 200)}

\item{max_avg_zyg_shift_neutral}{Maximum average zygosity shift for segments called as neutral before an adjustment is tagged as 'suspicious' (default 0.10)}

\item{min_avg_zyg_shift_onecopy}{Maximum difference between predicted value of zygosity shift for 1-copy regions and observed mean value (default 0.10)}
}
\value{
the same data.frame with an additional column named Suspicious taking values 0, 1, or 2.  1 means suspicious; 2 means very suspicious.
}
\description{
Label a set of adjustments that appear suspicious for biological reasons
}

