% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_functions.R
\name{m.step}
\alias{m.step}
\title{m.step}
\usage{
m.step(E.step.results, theta.seq, snplist, coveragelist, coverage_colname,
  stdev, copyrange = 0:4)
}
\arguments{
\item{E.step.results}{List representing output of e.step function}

\item{theta.seq}{Vector of tumor content values to consider}

\item{snplist}{List of snp data per segment, in the same order as segs data.frame}

\item{coveragelist}{List of coverage data per segment, in the same order as segs data.frame}

\item{coverage_colname}{Column name to use from coverage data.frame}

\item{stdev}{Global standard deviation of log2 coverage ratios}

\item{copyrange}{Vector of integer copy number states to consider (default 0:4)}
}
\value{
Value of theta (tumor purity) maximizing the likelihood
}
\description{
Perform M-Step of EM algorithm
}

