% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{make_germline_zygosity_plot}
\alias{make_germline_zygosity_plot}
\title{make_germline_zygosity_plot}
\usage{
make_germline_zygosity_plot(zygdata, polymorphic_regionsfile, outdir, prefix,
  chrinfo)
}
\arguments{
\item{zygdata}{Zygosity data.frame}

\item{polymorphic_regionsfile}{file containing polymorphic regions to exclude}

\item{outdir}{output directory to which to write plot (which goes to /germline subdirectory)}

\item{prefix}{sample prefix to use}

\item{chrinfo}{chromosome information data.frame}
}
\description{
Make a germline zygosity plot
}

