% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_functions.R
\name{merge_short_segments}
\alias{merge_short_segments}
\title{merge_short_segments}
\usage{
merge_short_segments(segdata)
}
\arguments{
\item{segdata}{input segmentation data.frame}
}
\value{
data.frame containing merged segments
}
\description{
Merge short segments of a segmentation data.frame: if a region has one non-missing exon and no snps, then merge into closest region
}

