% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-fit-functions.R
\name{models2fits}
\alias{models2fits}
\title{models2fits}
\usage{
models2fits(models, adjustments)
}
\arguments{
\item{models}{List of models}

\item{adjustments}{vector of adjustments corresponding to the models}
}
\value{
data.frame containing combined fits, ploidy, and adjustments for each model
}
\description{
Compute summary stats concerning goodness of fit for a set of models
}

