% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{multiComputeCoverage}
\alias{multiComputeCoverage}
\title{multiComputeCoverage}
\usage{
multiComputeCoverage(bamlist, bedfile, padding = 0, numproc = 2)
}
\arguments{
\item{bamlist}{list containing paths to bam files.  may have length 1 or 2.}

\item{bedfile}{bed file on which to compute coverage}

\item{padding}{number of bases (default 0) by which to pad targets in bed file}

\item{numproc}{number of processors to use (default 2)}
}
\value{
data.frame with columns Region, Chr, Start, End, Midpoint, Length, and either {Coverage} or {Coverage.T,Coverage.N} depending on length of bamlist
}
\description{
multiComputeCoverage
}

