% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_functions.R
\name{overall_classify}
\alias{overall_classify}
\title{overall_classify}
\usage{
overall_classify(Z.data, C.data, C.data.colname, seg.data, chrinfo,
  lower.limit.tc = 0.3, msg = "", exceptional_regions = NULL,
  gender = "female")
}
\arguments{
\item{Z.data}{zygosity data.frame}

\item{C.data}{coverage data.frame}

\item{C.data.colname}{Column name to use from coverage data.frame}

\item{seg.data}{data.frame containing segments}

\item{chrinfo}{data.frame containing chromosome info}

\item{lower.limit.tc}{Lower limit to consider for tumor content estimation (default = 0.3)}

\item{msg}{Message to print (default "")}

\item{exceptional_regions}{A set of regions that do not get classified (data.frame, columns Chr, Start, End. default NULL)}

\item{gender}{gender of sample (default female)}
}
\value{
a list with two components, classifs (data.frame of classifications) and tc (final inferred tumor content)
}
\description{
Overall classification function
}

