% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAR_coords_hg19}
\alias{PAR_coords_hg19}
\title{Pseudo-Autosomal Regions (hg19 coordinates)}
\format{A data frame}
\usage{
data(PAR_coords_hg19)
}
\description{
A data.frame containing hg19 coordinates of the pseudo-autosomal regions.
}
\keyword{datasets}

