% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_copynumber}
\alias{plot_copynumber}
\title{plot_copynumber}
\usage{
plot_copynumber(df_in, chrinfo, type = "coverage", coverage_colname = NULL,
  title_text = NULL, pointsize = 1)
}
\arguments{
\item{df_in}{input data.frame}

\item{chrinfo}{chromosome information data.frame}

\item{type}{'coverage' or 'zygosity'}

\item{coverage_colname}{column name to use for coverage}

\item{title_text}{text to use for the title of the plot}

\item{pointsize}{size of points in plot, default 1 (cex units)}
}
\description{
Make a copy number plot
}

