% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{prepare_coverage_data}
\alias{prepare_coverage_data}
\title{prepare_coverage_data}
\usage{
prepare_coverage_data(coverage_df, normalpool_file = NULL,
  pool_matched_gender = FALSE, add_plotting_coords = TRUE,
  polymorphic_regionsfile = NULL, numproc = 2, mincov = 30,
  gender = "infer", gender_inference_file = NULL)
}
\arguments{
\item{coverage_df}{data.frame of mean coverage quantifications (required)}

\item{normalpool_file}{filepath containing normal pool constructed with prepare_normal_pool function (optional)}

\item{pool_matched_gender}{if TRUE, pool has already been matched to the sample of interest.  default FALSE.}

\item{add_plotting_coords}{if TRUE, adds an extra column (Plot_Coord) containing coordinates for plotting.}

\item{numproc}{number of processors to use (default 2)}
}
\value{
data.frame containing GC-corrected coverage data of tumor (and optionally normal) sample against the normal sample and/or the normal pool
}
\description{
Prepare coverage data from bam files, a bed file, and a file containing quantifications across a normal pool
}

