% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{prepare_normal_pool}
\alias{prepare_normal_pool}
\title{prepare_normal_pool}
\usage{
prepare_normal_pool(pool_config_file, bedfile, outfile, numproc = 2,
  padding = 0)
}
\arguments{
\item{pool_config_file}{input tab-delimited text file (no header) with two columns containing paths to bam files and corresponding sample IDs, respectively}

\item{bedfile}{path to bed file containing target regions on which to compute coverage}

\item{outfile}{output file in which to save quantifications (.rds format)}

\item{numproc}{number of processors (default 2) to use}

\item{padding}{number of bases by which to extend targets on each side (default 0)}
}
\value{
data.frame containing pool data (targets x samples), which is also saved to outfile
}
\description{
Prepare a pool containing normal coverage quantifications
}

