% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{prepare_subread_annotation}
\alias{prepare_subread_annotation}
\title{prepare_subread_annotation}
\usage{
prepare_subread_annotation(bedfile, padding = 0)
}
\arguments{
\item{bedfile}{path to bedfile (can contain optional header, ignored if it exists.)}

\item{padding}{number of bases (default 0) on each side of target on which to compute coverage}
}
\value{
data.frame compatible with featureCounts function
}
\description{
prepare_subread_annotation
}

