% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygosity_functions.R
\name{vcf2zygosity}
\alias{vcf2zygosity}
\title{vcf2zygosity}
\usage{
vcf2zygosity(input_vcf, var_field = NULL, ref_field = NULL,
  tot_field = NULL, samplename_t = "Tumor", samplename_n = "Normal")
}
\arguments{
\item{input_vcf}{VCF file (may be gzipped) containing variants, variant IDs, and read counts}

\item{var_field}{Field containing per-variant variant read counts (e.g. "AO")}

\item{ref_field}{Field containing per-variant reference read counts (e.g. "RO")}

\item{tot_field}{Field containing per-variant total read counts (e.g. "DP")}

\item{samplename_t}{Sample name under which tumor sample appears (NULL if no tumor sample)}

\item{samplename_n}{Sample name under which normal sample appears (NULL if no normal sample)}
}
\value{
data.frame with columns ID, Chr, Pos, Ref, Alt, Ref.T, Var.T, Tot.T, Ref.N, Var.N, Tot.N
}
\description{
Convert a VCF file to a data.frame containing zygosity data for tumor/normal files
}

