% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygosity_functions.R
\name{vcf_to_df}
\alias{vcf_to_df}
\title{vcf_to_df}
\usage{
vcf_to_df(vcf)
}
\arguments{
\item{vcf}{Input VCF file (required)}
}
\value{
data.frame
}
\description{
Convert a VCF file to a data.frame
}

