% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygosity_functions.R
\name{vcf_to_popfreq_df}
\alias{vcf_to_popfreq_df}
\title{vcf_to_popfreq_df}
\usage{
vcf_to_popfreq_df(input_vcf, popfreq_field = "MAF")
}
\arguments{
\item{input_vcf}{input VCF file to parse}

\item{popfreq_field}{(default "MAF") INFO field containing the population frequency of interest}
}
\value{
data.frame with columns Chr, Pos, Ref, Alt, and Pop_Freq.  If popfreq_field is missing for a
variant, its population frequency is assumed to be zero.
}
\description{
Convert a VCF containing population frequencies into a data.frame
}

